Re-output query seqs. in last/train module, use them in last/lastal
Without this change, the channels containing query sequences and
alignment parameters become sorted differently, leading to wrong
parameters being used on wrong genomes.

In addition, the metadata is now conveyed in the `fastx` output
channel instead of the `param_file` one.  This removes one `map`
operation.  I checked separately that it was not the cause of
the problem.