Version 5.2.1 with more robust detection of index name.
Previously, the modules would search for a file ending in `.des` and
extract its base name, but on large genomes, the index is split in
multiple parts, and there is more than one `.des` file, causing
the command to crash.  This is solved by searching for a `.prj` file
instead, which is always unique.